+++ app/Http/Controllers/AnimalTypeController.php
@@ -16,11 +16,15 @@ class AnimalTypeController extends Controller
 
     public function create()
     {
+        $this->authorize('manage', AnimalType::class);
+
         return inertia('AnimalTypes/Create');
     }
 
     public function store(Request $request) 
     {
+        $this->authorize('manage', AnimalType::class);
+
         $request->validate(['name' => ['required']]);
 
         $created = AnimalType::create(['name' => $request->get('name')]);
@@ -30,6 +34,8 @@ class AnimalTypeController extends Controller
 
     public function destroy(AnimalType $animaltype)
     {
+        $this->authorize('manage', AnimalType::class);
+
         $animaltype->delete();
 
         return redirect()->route('animaltypes.index')->with('success', 'Állatfaj sikeresen törölve!');
